# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: CIBUS API Sistema de Gestión de Servicios Gastronómicos
  version: 1.0.0
  description: >
    Documentación de la API REST para el sistema de gestión de servicios gastronómicos CIBUS.
servers:
  - url: http://localhost:8080/api
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get users
      description: Returns a paginated list of users.
      parameters:
        - name: dni
          in: query
          schema:
            type: integer
        - name: name
          in: query
          schema:
            type: string
        - name: phoneNumber
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
        - name: userRoleId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'

    post:
      tags:
        - Users
      summary: Create user
      description: Creates a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'

    put:
      tags:
        - Users
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'

    delete:
      tags:
        - Users
      summary: Delete user
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204':
          description: User deleted successfully
  /users/{userId}/change-password:
    put:
      tags:
        - Users
      summary: Change user password
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDTO'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
  /users/{userId}/change-role:
    put:
      tags:
        - Users
      summary: Change user role
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRoleDTO'
      responses:
        '200':
          description: Role changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
  /users/{userId}/reset-pin:
    patch:
      tags:
        - Users
      summary: Reset user PIN
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPinRequestDTO'
      responses:
        '200':
          description: PIN reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
...
  /statistics/global/overview:
    get:
      tags:
        - Statistics
      summary: Get global overview insights
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Global overview statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewInsightDTO'
  /statistics/branch/{branchId}/overview:
    get:
      tags:
        - Statistics
      summary: Get branch overview insights
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Branch overview statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewInsightDTO'
  /statistics/branch/{branchId}/tables:
    get:
      tags:
        - Statistics
      summary: Get table insights
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Table statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableInsightDTO'
  /statistics/branch/{branchId}/products:
    get:
      tags:
        - Statistics
      summary: Get product insights for a branch
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Product statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInsightDTO'
  /statistics/global/products:
    get:
      tags:
        - Statistics
      summary: Get global product insights
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Global product statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInsightDTO'
  /statistics/branch/{branchId}/orders:
    get:
      tags:
        - Statistics
      summary: Get order insights for a branch
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Order statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInsightDTO'
  /statistics/global/orders:
    get:
      tags:
        - Statistics
      summary: Get global order insights
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Global order statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInsightDTO'
  /statistics/branch/{branchId}/waiters:
    get:
      tags:
        - Statistics
      summary: Get waiter insights for a branch
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Waiter statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaiterInsightDTO'
  /statistics/global/waiters:
    get:
      tags:
        - Statistics
      summary: Get global waiter insights
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Global waiter statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaiterInsightDTO'

# ── COMPONENTS ──────────────────────────────────────────────────────────────
components:
  parameters:
    branchId:
      name: branchId
      in: path
      required: true
      description: ID de la sucursal
      schema:
        type: integer
        example: 1
    userId:
      name: userId
      in: path
      required: true
      description: ID del usuario
      schema:
        type: integer
        example: 5
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    
  schemas:
...
    UserRequestDTO:
      type: object
      properties:
        dni:
          type: integer
          format: int64
          example: 40123456
        firstName:
          type: string
          example: "Juan"
        lastName:
          type: string
          example: "Pérez"
        phoneNumber:
          type: string
          example: "+5492235551234"
        email:
          type: string
          format: email
          example: "juan.perez@example.com"
        branchId:
          type: integer
          format: int64
          example: 1
        role:
          type: string
          example: "WAITER"
        pin:
          type: string
          example: "1234"
      required:
        - dni
        - firstName
        - lastName
        - phoneNumber
        - email
        - role
        - pin
    UserUpdateDTO:
      type: object
      properties:
        dni:
          type: integer
          format: int64
          example: 40123456
        firstName:
          type: string
          example: "Juan"
        lastName:
          type: string
          example: "Pérez"
        phoneNumber:
          type: string
          example: "+5492235551234"
        email:
          type: string
          format: email
          example: "juan.perez@example.com"
        branchId:
          type: integer
          format: int64
          example: 1
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        dni:
          type: integer
          format: int64
          example: 40123456
        firstName:
          type: string
          example: "Juan"
        lastName:
          type: string
          example: "Pérez"
        phoneNumber:
          type: string
          example: "+5492235551234"
        email:
          type: string
          format: email
          example: "juan.perez@example.com"
        branchId:
          type: integer
          format: int64
          example: 1
        branchName:
          type: string
          example: "Sucursal Centro"
        role:
          type: string
          example: "WAITER"
    ChangePasswordRequestDTO:
      type: object
      properties:
        currentPassword:
          type: string
          example: "oldPassword123"
        newPassword:
          type: string
          example: "newPassword456"
      required:
        - currentPassword
        - newPassword
    ChangeRoleDTO:
      type: object
      properties:
        roleName:
          type: string
          example: "MANAGER"
        newPin:
          type: string
          example: "4321"
      required:
        - roleName
    ResetPinRequestDTO:
      type: object
      properties:
        newPin:
          type: string
          example: "1234"
      required:
        - newPin

    # ── TABLE ────────────────────────────────────────────────────────────────
...
    OverviewInsightDTO:
      type: object
      properties:
        totalRevenue:
          type: number
          format: double
          example: 1250000.50
        totalOrders:
          type: integer
          format: int64
          example: 320
        occupancyRate:
          type: number
          format: double
          example: 82.5
        mostPopularProduct:
          $ref: '#/components/schemas/ProductMetricDTO'
        mostPopularCategory:
          $ref: '#/components/schemas/ProductCategoryMetricDTO'
        bestWaiter:
          $ref: '#/components/schemas/WaiterMetricDTO'
    TableInsightDTO:
      type: object
      properties:
        mostBusyTable:
          $ref: '#/components/schemas/TableMetricDTO'
        lessBusyTable:
          $ref: '#/components/schemas/TableMetricDTO'
        occupancyRate:
          type: number
          format: double
          example: 78.4
    ProductInsightDTO:
      type: object
      properties:
        mostPopularProduct:
          $ref: '#/components/schemas/ProductMetricDTO'
        lessPopularProduct:
          $ref: '#/components/schemas/ProductMetricDTO'
        mostExpensiveProduct:
          $ref: '#/components/schemas/ProductMetricDTO'
        cheapestProduct:
          $ref: '#/components/schemas/ProductMetricDTO'
        mostPopularCategory:
          $ref: '#/components/schemas/ProductCategoryMetricDTO'
        lessPopularCategory:
          $ref: '#/components/schemas/ProductCategoryMetricDTO'
    OrderInsightDTO:
      type: object
      properties:
        totalRevenue:
          type: number
          format: double
          example: 1250000.50
        averageTicket:
          type: number
          format: double
          example: 4200.75
        averageRevenuePerDay:
          type: number
          format: double
          example: 87500.20
        highestValueOrder:
          $ref: '#/components/schemas/OrderMetricDTO'
        lowestValueOrder:
          $ref: '#/components/schemas/OrderMetricDTO'
        totalOrders:
          type: integer
          format: int64
          example: 320
    WaiterInsightDTO:
      type: object
      properties:
        facturationRanking:
          type: array
          items:
            $ref: '#/components/schemas/WaiterMetricDTO'
        ordersQuantityRanking:
          type: array
          items:
            $ref: '#/components/schemas/WaiterMetricDTO'
        averageOrdersPerWaiter:
          type: number
          format: double
          example: 35.7
    ProductMetricDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 5
        productName:
          type: string
          example: "Milanesa Napolitana"
        productPrice:
          type: number
          format: double
          example: 4500.00
        timesSold:
          type: integer
          format: int64
          example: 350
    ProductCategoryMetricDTO:
      type: object
      properties:
        name:
          type: string
          example: "Main Course"
        timesSold:
          type: integer
          format: int64
          example: 1200
    WaiterMetricDTO:
      type: object
      properties:
        dni:
          type: integer
          format: int64
          example: 40123456
        firstName:
          type: string
          example: "Juan"
        lastName:
          type: string
          example: "Pérez"
        branchName:
          type: string
          example: "Sucursal Centro"
        facturation:
          type: number
          format: double
          example: 1542300.50
        ordersServed:
          type: integer
          format: int64
          example: 387
    TableMetricDTO:
      type: object
      properties:
        number:
          type: integer
          example: 12
        orderCount:
          type: integer
          format: int64
          example: 185
    OrderMetricDTO:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          example: 42
        branchId:
          type: integer
          format: int64
          example: 1
        tableId:
          type: integer
          format: int64
          example: 7
        waiterName:
          type: string
          example: "Juan Pérez"
        ticketValue:
          type: number
          format: double
          example: 15250.75
        orderDate:
          type: string
          format: date-time
          example: "2026-06-15T20:30:00-03:00"

    # ── ERROR ────────────────────────────────────────────────────────────────
...